requirements.txt
aiobotocore==1.0.7
asn1crypto==0.24.0
awesome-slugify==1.6.5
awscli==1.18.57
boto3==1.13.7
boto3-stubs==1.14.43.0
botocore==1.16.7
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cython==0.29.17
dateparser==0.7.4
Django==2.2.1
django-cognito-jwt==0.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.14
elasticsearch==7.8.0
email-validator==1.1.1
google-auth==1.14.3
gunicorn==19.9.0
idna==2.9
IPy==1.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
kubernetes==9.0.0
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
oauthlib==3.1.0
openapi-codec==1.3.2
pandas==0.24.2
pre-commit==2.4.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyreadstat==0.3.3
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2019.1
PyYAML==5.3.1
regex==2019.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3fs==0.5.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tzlocal==2.1
Unidecode==0.4.21
uritemplate==3.0.1
urllib3==1.25.9
websocket-client==0.56.0
